+++ src/main/java/com/wootecobook/turkey/login/controller/api/LoginApiControllerAdvice.java
@@ -1,8 +1,8 @@
 package com.wootecobook.turkey.login.controller.api;
 
+import com.wootecobook.turkey.commons.ErrorMessage;
 import com.wootecobook.turkey.commons.exception.NotLoginException;
 import com.wootecobook.turkey.login.service.exception.LoginFailException;
-import com.wootecobook.turkey.user.controller.api.ErrorMessage;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
