--- src/main/java/com/wootecobook/turkey/user/controller/api/ErrorMessage.java
@@ -1,16 +1,0 @@
-package com.wootecobook.turkey.user.controller.api;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class ErrorMessage {
-
-    private String errorMessage;
-
-    public ErrorMessage(String errorMessage) {
-        this.errorMessage = errorMessage;
-    }
-
-}
